requirements.txt
astor==0.8.1
attrdict==2.0.1
Babel==2.11.0
bce-python-sdk==0.8.74
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
contourpy==1.0.6
cryptography @ file:///croot/cryptography_1665612644927/work
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.6
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.38.0
future==0.18.2
gast==0.3.3
gevent==22.10.2
greenlet==2.0.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.23.0
imgaug==0.4.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Levenshtein==0.20.9
lmdb==1.4.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multiprocess==0.70.14
networkx==2.8.8
numpy==1.23.0
opencv-contrib-python==4.4.0.46
opencv-python==4.7.0.68
openpyxl==3.0.10
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==22.0
paddleocr==2.5.0.3
paddlepaddle-gpu @ file:///package/paddlepaddle_gpu-2.3.2-cp38-cp38-linux_x86_64.whl
pandas==1.5.2
Pillow==9.3.0
premailer==3.10.0
protobuf==3.19.1
pyclipper==1.3.0.post4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.16.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2022.7
PyWavelets==1.4.1
rapidfuzz==2.13.7
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-image==0.19.3
scipy==1.10.0
shapely==2.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
tifffile==2022.10.10
tqdm==4.64.1
urllib3 @ file:///croot/urllib3_1670526988650/work
visualdl==2.4.2
Werkzeug==2.2.2
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
